requirements.txt
Flask==1.1.2
gevent==20.9.0
keras==2.7.0
keras_nightly==2.5.0.dev2021032900
matplotlib==3.3.2
numpy==1.19.5
object_detection==0.0.3
Pillow==8.4.0
six==1.15.0
tensorflow==2.6.0
Werkzeug==1.0.1
